+++ CHANGELOG.md
@@ -8,6 +8,7 @@ PATCH
 * [Added] Added support for Welsh.
 
 ### Payments
+* [Added] Added API bindings for ACSS Debit.
 * [Added] Added API bindings for Scalapay.
 
 ### PaymentSheet